+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## Unreleased
 ## Added
 - Dynamic tick support for Preservation Disintegrate
+- Blacklist/Whitelist for cancelled spells on player bar
 
 ## Fixed
 - State refresh for spec swaps
+- Cancelled effect was appearing for hidden spells on player bar
 
 ---